Nettoyage des warning de conversion implicite de  type -en rendant ces dernières explicites. Modification des 'Screens' pour pouvoir modifier le goban depuis un menu (ce de chargement des problèmes en l'occurence).

+++ Jeu_de_Go/Sources/Graphics/Game/Board.h
@@ -70,7 +70,7 @@ Board::Board(const sf::Vector2<T>& _size) :
 		array[i] = new Square[size.y];
 		for (size_t j = 0; j < size.y; j++)
 		{
-			array[i][j].setPosition(sf::Vector2f(i, j));
+			array[i][j].setPosition(sf::Vector2f((float)i, (float)j));
 		}
 	}
 }
